+++ src/wave.rs
@@ -1,15 +1,15 @@
-use std::f64::consts::PI;
+use std::f32::consts::PI;
 use num_traits::sign::signum;
 
 /// add equations for waves here
-pub fn sine(freq: f64, time: f64, volume: f64 ) -> f32 {
-    return ((freq * time * PI * 2.).sin() * volume) as f32;
+pub fn sine(freq: f32, time: f32, volume: f32 ) -> f32 {
+    return (freq * time * PI * 2.).sin() * volume;
 }
 
-pub fn square(freq: f64, time: f64, volume: f64 ) -> f32 {
-    return (signum((freq * time * PI * 2.).sin()) * volume) as f32;
+pub fn square(freq: f32, time: f32, volume: f32 ) -> f32 {
+    return signum((freq * time * PI * 2.).sin()) * volume;
 }
 
-pub fn saw(freq: f64, time: f64, volume: f64) -> f32 {
-    return (((freq/2.) * time) % volume) as f32
+pub fn saw(freq: f32, time: f32, volume: f32) -> f32 {
+    return ((freq/2.) * time) % volume
 }
